requirements.txt
langchain-openai>=0.0.2
langchain-core>=0.1.0
openai>=1.0.0
pandas>=1.3.0
pytesseract>=0.3.8
pdf2image>=1.16.0
python-dotenv>=0.19.0
ocrmypdf>=14.0.0
docx2txt>=0.8